Create page1.html
I had to break down the page into two sections since there is a ruling with how many apis I can have in one html file. So I will have the user start off with page1.html and then they should be able to move to page2.html when they want to see the other data. It is not what I wanted, but I need to make this work for this upcoming week.